loongarch: Add basic support for LoongArch32
